requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.19
email-validator==2.2.0
